- Move FoundGames into MyGames view. - Update layout to new design
