--- src/NexusMods.App.UI/Controls/FoundGames/FoundGamesView.axaml
@@ -1,36 +1,0 @@
-﻿<reactiveUi:ReactiveUserControl
-    d:DesignHeight="400"
-    d:DesignWidth="896"
-    mc:Ignorable="d"
-    x:Class="NexusMods.App.UI.Controls.FoundGames.FoundGamesView"
-    x:TypeArguments="foundGames:IFoundGamesViewModel"
-    xmlns="https://github.com/avaloniaui"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:reactiveUi="http://reactiveui.net"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:foundGames="clr-namespace:NexusMods.App.UI.Controls.FoundGames">
-    <Design.DataContext>
-        <foundGames:FoundGamesDesignViewModel />
-    </Design.DataContext>
-    <Border Classes="Low Rounded-lg" Padding="24">
-        <ItemsControl
-            x:Name="FoundGamesItemsControl">
-            <ItemsControl.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <WrapPanel />
-                </ItemsPanelTemplate>
-            </ItemsControl.ItemsPanel>
-            <ItemsControl.ItemTemplate>
-                <DataTemplate>
-                    <reactiveUi:ViewModelViewHost
-                        Height="238"
-                        Margin="0,0,0,24"
-                        ViewModel="{Binding .}"
-                        Width="192" />
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
-    </Border>
-
-</reactiveUi:ReactiveUserControl>
