+++ src/NexusMods.App.UI/Pages/MyGames/IMyGamesViewModel.cs
@@ -1,12 +1,14 @@
-using NexusMods.App.UI.Controls.FoundGames;
+using System.Collections.ObjectModel;
+using NexusMods.App.UI.Controls.GameWidget;
 using NexusMods.App.UI.WorkspaceSystem;
 
 namespace NexusMods.App.UI.Pages.MyGames;
 
 public interface IMyGamesViewModel : IPageViewModelInterface
 {
-    public IFoundGamesViewModel FoundGames { get; }
-
-    public IFoundGamesViewModel AllGames { get; }
+    
+    public ReadOnlyObservableCollection<IGameWidgetViewModel> ManagedGames { get; }
+    
+    public ReadOnlyObservableCollection<IGameWidgetViewModel> DetectedGames { get; }
 
 }
