+++ src/NexusMods.App.UI/Pages/MyGames/MyGamesViewModel.cs
@@ -1,6 +1,12 @@
+using System.Collections.ObjectModel;
+using Avalonia.Threading;
 using JetBrains.Annotations;
+using Microsoft.Extensions.DependencyInjection;
+using NexusMods.Abstractions.GameLocators;
 using NexusMods.Abstractions.Games;
-using NexusMods.App.UI.Controls.FoundGames;
+using NexusMods.Abstractions.Loadouts;
+using NexusMods.App.UI.Controls.GameWidget;
+using NexusMods.App.UI.Pages.LoadoutGrid;
 using NexusMods.App.UI.Resources;
 using NexusMods.App.UI.Windows;
 using NexusMods.App.UI.WorkspaceSystem;
@@ -12,22 +18,27 @@ namespace NexusMods.App.UI.Pages.MyGames;
 [UsedImplicitly]
 public class MyGamesViewModel : APageViewModel<IMyGamesViewModel>, IMyGamesViewModel
 {
-    [Reactive] public IFoundGamesViewModel FoundGames { get; set; }
-
-    [Reactive] public IFoundGamesViewModel AllGames { get; set; }
+    private readonly ILoadoutRegistry _loadoutRegistry;
+    private readonly IServiceProvider _provider;
+    private readonly IWindowManager _windowManager;
+    
+    [Reactive] public ReadOnlyObservableCollection<IGameWidgetViewModel> ManagedGames { get; private set; }
+    public ReadOnlyObservableCollection<IGameWidgetViewModel> DetectedGames { get; private set; }
 
     public MyGamesViewModel(
         IWindowManager windowManager,
-        IFoundGamesViewModel foundGames,
-        IFoundGamesViewModel allGames,
+        IServiceProvider serviceProvider,
+        ILoadoutRegistry loadoutRegistry,
         IEnumerable<IGame> games) : base(windowManager)
     {
+        _provider = serviceProvider;
+        _loadoutRegistry = loadoutRegistry;
+        _windowManager = windowManager;
+        
         var gamesList = games.ToList();
-        FoundGames = foundGames;
-        foundGames.InitializeFromFound(gamesList);
-
-        AllGames = allGames;
-        allGames.InitializeManual(gamesList);
+        DetectedGames = GetDetectedGames(gamesList);
+        // TODO: Implement ManagedGames
+        ManagedGames = GetDetectedGames(gamesList);
 
         this.WhenActivated(() =>
         {
@@ -36,4 +47,53 @@ public class MyGamesViewModel : APageViewModel<IMyGamesViewModel>, IMyGamesViewM
             return Array.Empty<IDisposable>();
         });
     }
+    
+    private async Task ManageGame(GameInstallation installation)
+    {
+        var name = _loadoutRegistry.SuggestName(installation);
+        var marker = await _loadoutRegistry.Manage(installation, name);
+
+        var loadoutId = marker.Id;
+
+        Dispatcher.UIThread.Invoke(() =>
+        {
+            if (!_windowManager.TryGetActiveWindow(out var window)) return;
+            var workspaceController = window.WorkspaceController;
+
+            workspaceController.ChangeOrCreateWorkspaceByContext(
+                context => context.LoadoutId == loadoutId,
+                () => new PageData
+                {
+                    FactoryId = LoadoutGridPageFactory.StaticId,
+                    Context = new LoadoutGridContext
+                    {
+                        LoadoutId = loadoutId
+                    }
+                },
+                () => new LoadoutContext
+                {
+                    LoadoutId = loadoutId
+                }
+            );
+        });
+    }
+    
+    private ReadOnlyObservableCollection<IGameWidgetViewModel> GetDetectedGames(IEnumerable<IGame> games)
+    {
+        var installed = games
+            .SelectMany(g => g.Installations)
+            .Select(install =>
+            {
+                var vm = _provider.GetRequiredService<IGameWidgetViewModel>();
+                vm.Installation = install;
+                vm.PrimaryButton = ReactiveCommand.CreateFromTask(async () =>
+                {
+                    await Task.Run(async () => await ManageGame(install));
+                });
+                return vm;
+            });
+        return new ReadOnlyObservableCollection<IGameWidgetViewModel>(new ObservableCollection<IGameWidgetViewModel>(installed));
+    }
+
+
 }
